requirements.txt
streamlit
openai
python-dotenv
requests
pyyaml
langchain
langchain-community
chromadb
pysqlite3-binary
onnxruntime
pandas
numpy
nb-clean
ipykernel
google-cloud-aiplatform
google-genai
google-cloud-logging
